postfix "i64" is invalid in gcc.

+++ src/environ.cc
@@ -400,7 +400,7 @@ decoded_time_to_universal_time (int year, int mon, int day,
 }
 
 #define FILETIME_UNIT_PER_SECOND 10000000
-#define FILETIME_UTC_BASE   9435484800i64   // FileTime (1900/1/1 0:0:0)
+#define FILETIME_UTC_BASE (static_cast <int64_t> (9435484800))   // FileTime (1900/1/1 0:0:0)
 
 lisp
 file_time_to_universal_time (const FILETIME &ft)
